#22 - feat : SignInState를 하나의 클래스파일로 분리
